requirements.txt
streamlit
pandas
numpy
easyocr
opencv-python
matplotlib
google-generativeai
python-dotenv
jupyter
openpyxl